requirements.txt
# Core dependencies
akshare>=1.16.94
langchain>=0.1.0
ollama>=0.1.0
# Using qwen3-nothink:14b (custom model with thinking mode disabled)
numpy>=1.21.0
pandas>=1.3.0
pydantic>=2.0.0

# Web interface dependencies
flask>=2.0.0
matplotlib>=3.4.0

# Visualization dependencies
plotly>=5.0.0

# Utility dependencies
python-dotenv>=1.0.0
requests>=2.25.0
